Only call scionlab-config in postinst if scionlab-config.json exists
It seems like a likely scenario that people first install the packages and *then* put the configuration files in place.
If scionlab-config.json doesn't exist, scionlab-config will only complain that it doesn't have the required host-id/secret and not do anything else.
Avoid confusing people with an error message in a situation that is completely fine.

Also, always call `scionlab-config` with `--force`, as it avoids the branch and I can't think of a good enough reason not to.