Avoid error on startup with non-existing sciond-dir
If the directory `/run/shm/sciond` does not exist, the `rm` command in `ExecStartPre` would fail. This does not prevent sciond from starting properly, but it shows up as a huge red `(code=exited, status=1/FAILURE)` message in `systemctl status`. This is just confusing, as it's not obvious that this can safely be ignored.

Prevent the confusing error message by making `rm` ignore missing files.