Deploy staging packages to separate test repository
The previous approach with a test distribution was severly flawed --
Packages are shared between the different distributinos! Once a package
with a certain name/version was deployed, e.g. on a test branch, it
would never be overridden by the actual deployment.

Also drop the preprod stage (deployed packages to a "preprod"
distribution before actually deploying them to "all"), for the same reason.
The preprod also did not seem to serve any actual purpose, as there is
no clear case in which this would catch an actual problem before deployment.